更新社会化账号会员注册机制 更新系统图形验证码样式 修复用户阅读全选bug 修复静态页生成进度显示错误bug 修复使用sqlite数据库不能删除网站多语言的bug 修复可视化管理员权限控制失效的bug 修复sql注入漏洞 修复产品所属多个栏目时，管理后台部分栏目列表下不显示的问题 修复开启商城后新增产品演示视频不显示的问题 修复前台模板加载js没有加载浏览器缓存的问题 修复markdown编辑器中有表格时，保存提示错误信息的问题 修复手机端详情页tag词过长时不换行的问题 修复商品详情编辑时详情选项卡跟系统设置不一致的问题 系统表格组件更新 修复系统xss漏洞 修复备份数据导入可能意外终止的问题 修复部分模板演示视频不显示的问题 优化系统代码

+++ app/system/language/admin/language_web.class.php
@@ -242,23 +242,19 @@ class language_web extends admin
         }
 
         //替换数据表里带语言参数的数据
-        $tables = DB::get_all('show tables');
-        foreach ($tables as $key => $value) {
-            foreach ($value as $table_name) {
-                if (!$table_name) {
-                    continue;
-                }
-                $fields = DB::get_all("desc " . $table_name);
-                foreach ($fields as $val) {
-                    if ($val['Field'] == 'lang') {
-                        $query = "DELETE FROM {$table_name} WHERE lang ='{$lang_info['lang']}'";
-                        if (strstr($table_name, 'language')) {
-                            $query .= " and site ='0'";
-                        }
-                        DB::query($query);
-                    }
-                }
+        $excepted = array(
+            "lang_admin",
+        );
+        foreach ($_M['table'] as $key => $table_name) {
+            if (!$table_name || in_array($key,$excepted)) {
+                continue;
             }
+
+            $query = "DELETE FROM {$table_name} WHERE lang ='{$lang_info['lang']}'";
+            if ($table_name == "{$_M['config']['tablepre']}language") {
+                $query .= " AND site ='0'";
+            }
+            DB::query($query);
         }
 
         $query = "SELECT id FROM {$_M['table']['lang']} order by no_order where lang!='metinfo'";
